fix: move JUnit 5 tests to separate class
